+++ contracts/token/ERC1155/IERC1155MetadataURI.sol
@@ -1,19 +1,21 @@
-pragma solidity ^0.6.6;
+// SPDX-License-Identifier: MIT
+
+pragma solidity 0.6.8;
 
 /**
-    @title ERC-1155 Multi Token Standard, optional metadata URI extension
-    @dev See https://eips.ethereum.org/EIPS/eip-1155
-    Note: The ERC-165 identifier for this interface is 0x0e89341c.
+ * @title ERC-1155 Multi Token Standard, optional metadata URI extension
+ * @dev See https://eips.ethereum.org/EIPS/eip-1155
+ * Note: The ERC-165 identifier for this interface is 0x0e89341c.
  */
 interface IERC1155MetadataURI {
     /**
-        @notice A distinct Uniform Resource Identifier (URI) for a given token.
-        @dev URIs are defined in RFC 3986.
-        The URI MUST point to a JSON file that conforms to the "ERC-1155 Metadata URI JSON Schema".
-        The uri function SHOULD be used to retrieve values if no event was emitted.
-        The uri function MUST return the same value as the latest event for an _id if it was emitted.
-        The uri function MUST NOT be used to check for the existence of a token as it is possible for an implementation to return a valid string even if the token does not exist.
-        @return URI string
-    */
+     * @notice A distinct Uniform Resource Identifier (URI) for a given token.
+     * @dev URIs are defined in RFC 3986.
+     * The URI MUST point to a JSON file that conforms to the "ERC-1155 Metadata URI JSON Schema".
+     * The uri function SHOULD be used to retrieve values if no event was emitted.
+     * The uri function MUST return the same value as the latest event for an _id if it was emitted.
+     * The uri function MUST NOT be used to check for the existence of a token as it is possible for an implementation to return a valid string even if the token does not exist.
+     * @return URI string
+     */
     function uri(uint256 id) external view returns (string memory);
 }