+++ migrations/3_NonBurnableInventory.js
@@ -1,5 +1,0 @@
-const NonBurnableInventoryMock = artifacts.require('NonBurnableInventoryMock');
-
-module.exports = async (deployer, network, accounts) => { 
-    await deployer.deploy(NonBurnableInventoryMock, 1);
-}