+++ src/constants.js
@@ -21,8 +21,25 @@ const ERC1155AssetCollections_InterfaceId_Experimental = interfaces.ERC1155Asset
 const ERC1155MetadataURI_InterfaceId = interfaces.ERC1155MetadataURI.id;
 const ERC1155TokenReceiver_InterfaceId = interfaces.ERC1155TokenReceiver;
 
+const DefaultNFMaskLength = 32;
+
+const DefaultFungibleLayout = [
+    { name: 'baseCollectionId', bits: 256 },
+];
+
+const DefaultNonFungibleLayout = [
+    { name: 'baseTokenId', bits: 256 - DefaultNFMaskLength },
+    { name: 'baseCollectionId', bits: DefaultNFMaskLength - 1 },
+    { name: 'nfFlag', bits: 1 }
+];
 
 module.exports = {
+    // Number
+    DefaultNFMaskLength,
+    Number: {
+        DefaultNFMaskLength
+    },
+
     // Bytes4
     ERC721Received_MagicValue, ERC1155Received_MagicValue, ERC1155BatchReceived_MagicValue,
     ERC721_InterfaceId, ERC721Metadata_InterfaceId, ERC721Enumerable_InterfaceId, ERC721Exists_InterfaceId_Experimental, ERC721Receiver_InterfaceId,
@@ -31,5 +48,8 @@ module.exports = {
         ERC721Received_MagicValue, ERC1155Received_MagicValue, ERC1155BatchReceived_MagicValue,
         ERC721_InterfaceId, ERC721Metadata_InterfaceId, ERC721Enumerable_InterfaceId, ERC721Exists_InterfaceId_Experimental, ERC721Receiver_InterfaceId,
         ERC1155_InterfaceId, ERC1155MetadataURI_InterfaceId, ERC1155AssetCollections_InterfaceId_Experimental, ERC1155TokenReceiver_InterfaceId,
-    }
+    },
+
+    // Bits Layouts
+    DefaultFungibleLayout, DefaultNonFungibleLayout,
 };