+++ src/index.js
@@ -1,7 +1,9 @@
 const constants = require('./constants');
 const interfaces = require('./interfaces/ERC165');
+const bytes32Attributes = require('./helpers/bytes32Attributes');
 
 module.exports = {
     constants,
-    interfaces
+    interfaces,
+    bytes32Attributes
 }